change text for hideBtn
