+++ js/cardRender.js
@@ -1,5 +1,6 @@
 
 import DoctorAPIService from './doctor_api_service.js';
+import toggleHideParameters from './HideParameters.js';
 const request = new DoctorAPIService()
 let visitCards = document.querySelectorAll('.visit-wrap .visit-card');
 const list = document.querySelector('.visit-wrap');
@@ -8,19 +9,7 @@ const selectStatus = document.querySelector('.select-options');
 const selectUrgency = document.querySelector('.select-urgency');
 
 
-function toggleHideParameters(e) {
-   const wrap = e.target.closest('.visit-card').querySelector('.parameters-wrap');
-   const parametersWrap = e.target.closest('.visit-card').querySelector('.hide-parameters');
-   const hideAge = e.target.closest('.visit-card').querySelector('.hide-age');
-   hideAge.classList.toggle('hide')
-   parametersWrap.classList.toggle('hide');
-   wrap.classList.toggle('hide');
-   if (!wrap.classList.contains('hide')) {
-      e.target.textContent = 'Показать меньше'
-   } else {
-      e.target.textContent = 'Показать больше'
-   }
-}
+
 
 visitCards.forEach((card) => {
    card.addEventListener('click', (e) => {
